+++ neo4j/datadog_checks/neo4j/__init__.py
@@ -1,3 +1,8 @@
+import sys
+
+if not sys.executable:
+    sys.executable = None
+
 from .__about__ import __version__
 from .neo4j import Neo4jCheck