+++ neo4j/datadog_checks/neo4j/metrics.py
@@ -11,166 +11,192 @@ METRIC_MAP = {
     'bolt_messages_received': 'bolt.messages_received',
     'bolt_messages_started': 'bolt.messages_started',
     # causal clustering metrics
-    'causal_clustering_catchup_tx_pull_requests_received': 'causal_clustering.catchup_tx_pull_requests_received',
-    'causal_clustering_core_append_index': 'causal_clustering.core.append_index',
-    'causal_clustering_core_commit_index': 'causal_clustering.core.commit_index',
-    'causal_clustering_core_discovery_cluster_converged': 'causal_clustering.core.discovery.cluster.converged',
-    'causal_clustering_core_discovery_cluster_members': 'causal_clustering.core.discovery.cluster.members',
-    'causal_clustering_core_discovery_cluster_unreachable': 'causal_clustering.core.discovery.cluster.unreachable',
+    'cluster_catchup_tx_pull_requests_received': 'cluster.catchup_tx_pull_requests_received',
+    'cluster_core_append_index': 'cluster.core.append_index',
+    'cluster_core_commit_index': 'cluster.core.commit_index',
+    'cluster_core_discovery_cluster_converged': 'cluster.core.discovery.cluster.converged',
+    'cluster_core_discovery_cluster_members': 'cluster.core.discovery.cluster.members',
+    'cluster_core_discovery_cluster_unreachable': 'cluster.core.discovery.cluster.unreachable',
     # causal clustering replicated data metrics (4.3)
-    'causal_clustering_core_discovery_replicated_data_member_data_invisible': (
-        'causal_clustering.core.discovery.replicated_data.member_data.invisible'
+    'cluster_core_discovery_replicated_data_member_data_invisible': (
+        'cluster.core.discovery.replicated_data.member_data.invisible'
     ),
-    'causal_clustering_core_discovery_replicated_data_member_data_visible': (
-        'causal_clustering.core.discovery.replicated_data.member_data.visible'
+    'cluster_core_discovery_replicated_data_member_data_visible': (
+        'cluster.core.discovery.replicated_data.member_data.visible'
     ),
-    'causal_clustering_core_discovery_replicated_data_raft_id_published_by_member_invisible': (
-        'causal_clustering.core.discovery.replicated_data.raft_id_published_by_member.invisible'
+    'cluster_core_discovery_replicated_data_raft_id_published_by_member_invisible': (
+        'cluster.core.discovery.replicated_data.raft_id_published_by_member.invisible'
     ),
-    'causal_clustering_core_discovery_replicated_data_raft_id_published_by_member_visible': (
-        'causal_clustering.core.discovery.replicated_data.raft_id_published_by_member.visible'
+    'cluster_core_discovery_replicated_data_raft_id_published_by_member_visible': (
+        'cluster.core.discovery.replicated_data.raft_id_published_by_member.visible'
     ),
-    'causal_clustering_core_discovery_replicated_data_per_db_leader_name_invisible': (
-        'causal_clustering.core.discovery.replicated_data.per_db_leader_name.invisible'
+    'cluster_core_discovery_replicated_data_per_db_leader_name_invisible': (
+        'cluster.core.discovery.replicated_data.per_db_leader_name.invisible'
     ),
-    'causal_clustering_core_discovery_replicated_data_per_db_leader_name_visible': (
-        'causal_clustering.core.discovery.replicated_data.per_db_leader_name.visible'
+    'cluster_core_discovery_replicated_data_per_db_leader_name_visible': (
+        'cluster.core.discovery.replicated_data.per_db_leader_name.visible'
     ),
-    'causal_clustering_core_discovery_replicated_data_member_db_state_invisible': (
-        'causal_clustering.core.discovery.replicated_data.member_db_state.invisible'
+    'cluster_core_discovery_replicated_data_member_db_state_invisible': (
+        'cluster.core.discovery.replicated_data.member_db_state.invisible'
     ),
-    'causal_clustering_core_discovery_replicated_data_member_db_state_visible': (
-        'causal_clustering.core.discovery.replicated_data.member_db_state.visible'
+    'cluster_core_discovery_replicated_data_member_db_state_visible': (
+        'cluster.core.discovery.replicated_data.member_db_state.visible'
     ),
-    'causal_clustering_core_discovery_replicated_data_raft_member_mapping_invisible': (
-        'causal_clustering.core.discovery.replicated_data.raft_member_mapping.invisible'
+    'cluster_core_discovery_replicated_data_raft_member_mapping_invisible': (
+        'cluster.core.discovery.replicated_data.raft_member_mapping.invisible'
     ),
-    'causal_clustering_core_discovery_replicated_data_raft_member_mapping_visible': (
-        'causal_clustering.core.discovery.replicated_data.raft_member_mapping.visible'
+    'cluster_core_discovery_replicated_data_raft_member_mapping_visible': (
+        'cluster.core.discovery.replicated_data.raft_member_mapping.visible'
     ),
     # causal clustering replicated data metrics (3.6)
-    'causal_clustering_core_discovery_replicated_data_cluster_id_per_db_name_invisible': (
-        'causal_clustering.core.discovery.replicated_data.cluster_id_per_db_name.invisible'
+    'cluster_core_discovery_replicated_data_cluster_id_per_db_name_invisible': (
+        'cluster.core.discovery.replicated_data.cluster_id_per_db_name.invisible'
     ),
-    'causal_clustering_core_discovery_replicated_data_cluster_id_per_db_name_visible': (
-        'causal_clustering.core.discovery.replicated_data.cluster_id_per_db_name.visible'
+    'cluster_core_discovery_replicated_data_cluster_id_per_db_name_visible': (
+        'cluster.core.discovery.replicated_data.cluster_id_per_db_name.visible'
     ),
     # end causal clustering replicated data metrics
-    'causal_clustering_core_in_flight_cache_element_count': 'causal_clustering.core.in_flight_cache_element_count',
-    'causal_clustering_core_in_flight_cache_hits': 'causal_clustering.core.in_flight_cache.hits',
-    'causal_clustering_core_in_flight_cache_max_bytes': 'causal_clustering.core.in_flight_cache.max_bytes',
-    'causal_clustering_core_in_flight_cache_max_elements': 'causal_clustering.core.in_flight_cache.max_elements',
-    'causal_clustering_core_in_flight_cache_misses': 'causal_clustering.core.in_flight_cache.misses',
-    'causal_clustering_core_in_flight_cache_total_bytes': 'causal_clustering.core.in_flight_cache.total_bytes',
-    'causal_clustering_core_is_leader': 'causal_clustering.core.is_leader',
-    'causal_clustering_core_message_processing_delay': 'causal_clustering.core.message_processing_delay',
-    'causal_clustering_core_message_processing_timer': 'causal_clustering.core.message_processing_timer',
-    'causal_clustering_core_message_processing_timer_append_entries_request': (
-        'causal_clustering.core.message_processing_timer.append_entries_request'
+    'cluster_core_in_flight_cache_element_count': 'cluster.core.in_flight_cache_element_count',
+    'cluster_core_in_flight_cache_hits': 'cluster.core.in_flight_cache.hits',
+    'cluster_core_in_flight_cache_max_bytes': 'cluster.core.in_flight_cache.max_bytes',
+    'cluster_core_in_flight_cache_max_elements': 'cluster.core.in_flight_cache.max_elements',
+    'cluster_core_in_flight_cache_misses': 'cluster.core.in_flight_cache.misses',
+    'cluster_core_in_flight_cache_total_bytes': 'cluster.core.in_flight_cache.total_bytes',
+    'cluster_core_is_leader': 'cluster.core.is_leader',
+    'cluster_core_message_processing_delay': 'cluster.core.message_processing_delay',
+    'cluster_core_message_processing_timer': 'cluster.core.message_processing_timer',
+    'cluster_core_message_processing_timer_append_entries_request': (
+        'cluster.core.message_processing_timer.append_entries_request'
     ),
-    'causal_clustering_core_message_processing_timer_append_entries_response': (
-        'causal_clustering.core.message_processing_timer.append_entries_response'
+    'cluster_core_message_processing_timer_append_entries_response': (
+        'cluster.core.message_processing_timer.append_entries_response'
     ),
-    'causal_clustering_core_message_processing_timer_election_timeout': (
-        'causal_clustering.core.message_processing_timer.election_timeout'
+    'cluster_core_message_processing_timer_election_timeout': (
+        'cluster.core.message_processing_timer.election_timeout'
     ),
-    'causal_clustering_core_message_processing_timer_heartbeat': (
-        'causal_clustering.core.message_processing_timer.heartbeat'
+    'cluster_core_message_processing_timer_heartbeat': ('cluster.core.message_processing_timer.heartbeat'),
+    'cluster_core_message_processing_timer_heartbeat_response': (
+        'cluster.core.message_processing_timer.heartbeat_response'
     ),
-    'causal_clustering_core_message_processing_timer_heartbeat_response': (
-        'causal_clustering.core.message_processing_timer.heartbeat_response'
+    'cluster_core_message_processing_timer_heartbeat_timeout': (
+        'cluster.core.message_processing_timer.heartbeat_timeout'
     ),
-    'causal_clustering_core_message_processing_timer_heartbeat_timeout': (
-        'causal_clustering.core.message_processing_timer.heartbeat_timeout'
+    'cluster_core_message_processing_timer_log_compaction_info': (
+        'cluster.core.message_processing_timer.log_compaction_info'
     ),
-    'causal_clustering_core_message_processing_timer_log_compaction_info': (
-        'causal_clustering.core.message_processing_timer.log_compaction_info'
+    'cluster_core_message_processing_timer_new_batch_request': (
+        'cluster.core.message_processing_timer.new_batch_request'
     ),
-    'causal_clustering_core_message_processing_timer_new_batch_request': (
-        'causal_clustering.core.message_processing_timer.new_batch_request'
+    'cluster_core_message_processing_timer_new_entry_request': (
+        'cluster.core.message_processing_timer.new_entry_request'
     ),
-    'causal_clustering_core_message_processing_timer_new_entry_request': (
-        'causal_clustering.core.message_processing_timer.new_entry_request'
+    'cluster_core_message_processing_timer_pre_vote_request': (
+        'cluster.core.message_processing_timer.pre_vote_request'
     ),
-    'causal_clustering_core_message_processing_timer_pre_vote_request': (
-        'causal_clustering.core.message_processing_timer.pre_vote_request'
+    'cluster_core_message_processing_timer_pre_vote_response': (
+        'cluster.core.message_processing_timer.pre_vote_response'
     ),
-    'causal_clustering_core_message_processing_timer_pre_vote_response': (
-        'causal_clustering.core.message_processing_timer.pre_vote_response'
+    'cluster_core_message_processing_timer_prune_request': ('cluster.core.message_processing_timer.prune_request'),
+    'cluster_core_message_processing_timer_vote_request': ('cluster.core.message_processing_timer.vote_request'),
+    'cluster_core_message_processing_timer_vote_response': ('cluster.core.message_processing_timer.vote_response'),
+    'cluster.read_replica.pull_updates': 'cluster.read_replica.pull_updates',
+    'cluster.read_replica.pull_update_highest_tx_id_requested': (
+        'cluster.read_replica.pull_update_highest_tx_id_requested'
     ),
-    'causal_clustering_core_message_processing_timer_prune_request': (
-        'causal_clustering.core.message_processing_timer.prune_request'
+    'cluster.read_replica.pull_update_highest_tx_id_received': (
+        'cluster.read_replica.pull_update_highest_tx_id_received'
     ),
-    'causal_clustering_core_message_processing_timer_vote_request': (
-        'causal_clustering.core.message_processing_timer.vote_request'
-    ),
-    'causal_clustering_core_message_processing_timer_vote_response': (
-        'causal_clustering.core.message_processing_timer.vote_response'
-    ),
-    'causal_clustering.read_replica.pull_updates': 'causal_clustering.read_replica.pull_updates',
-    'causal_clustering.read_replica.pull_update_highest_tx_id_requested': (
-        'causal_clustering.read_replica.pull_update_highest_tx_id_requested'
-    ),
-    'causal_clustering.read_replica.pull_update_highest_tx_id_received': (
-        'causal_clustering.read_replica.pull_update_highest_tx_id_received'
-    ),
-    'causal_clustering_core_replication_attempt': 'causal_clustering.core.replication_attempt',
-    'causal_clustering_core_replication_fail': 'causal_clustering.core.replication_fail',
-    'causal_clustering_core_replication_new': 'causal_clustering.core.replication_new',
-    'causal_clustering_core_replication_success': 'causal_clustering.core.replication_success',
-    'causal_clustering_core_term': 'causal_clustering.core.term',
-    'causal_clustering_core_tx_retries': 'causal_clustering.core.tx_retries',
+    'cluster_core_replication_attempt': 'cluster.core.replication_attempt',
+    'cluster_core_replication_fail': 'cluster.core.replication_fail',
+    'cluster_core_replication_new': 'cluster.core.replication_new',
+    'cluster_core_replication_success': 'cluster.core.replication_success',
+    'cluster_core_term': 'cluster.core.term',
+    'cluster_core_tx_retries': 'cluster.core.tx_retries',
     # database checkpointing metrics
     'check_point_events': 'check_point.events',
     'check_point_total_time': 'check_point.total_time',
     'check_point_duration': 'check_point.duration',
+    'check_point_pages_flushed': 'check_point.pages_flushed',
+    'check_point_io_performed': 'check_point.io_performed',
+    'check_point_io_limit': 'check_point.io_limit',
     # cypher metrics
     'cypher_replan_events': 'cypher.replan_events',
     'cypher_replan_wait_time': 'cypher.replan_wait_time',
-    # database data metrics
+    'cypher_cache_ast_entries': 'cypher.cache.ast.entries',
+    'cypher_cache_executable_query_entries': 'cypher.cache.executable_query.entries',
+    'cypher_cache_execution_plan_entries': 'cypher.cache.execution_plan.entries',
+    'cypher_cache_logical_plan_entries': 'cypher.cache.logical_plan.entries',
+    'cypher_cache_pre_parser_entries': 'cypher.cache.pre_parser.entries',
+    # database data count metrics
     'neo4j_count_node': 'node_count',
     'neo4j_count_relationship': 'relationship_count',
+    # database data metrics
     'ids_in_use_node': 'ids_in_use.node',
     'ids_in_use_property': 'ids_in_use.property',
     'ids_in_use_relationship': 'ids_in_use.relationship',
     'ids_in_use_relationship_type': 'ids_in_use.relationship_type',
-    'store_size_total': 'store.size.total',
-    'store_size_database': 'store.size.database',
+    # database operation count metrics
+    'db_operations_count_create': 'db_operation_count.create',
+    'db_operations_count_start': 'db_operation_count.start',
+    'db_operations_count_stop': 'db_operation_count.stop',
+    'db_operations_count_drop': 'db_operation_count.drop',
+    'db_operations_count_failed': 'db_operation_count.failed',
+    'db_operations_count_recovered': 'db_operation_count.recovered',
+    # database store size metrics
+    'neo4j_store_size_total': 'neo4j.store.size_total',
+    'neo4j_store_size_database': 'neo4j.store.size_database',
     # database transaction log metrics
     'log_appended_bytes': 'log.appended_bytes',
+    'log_append_batch_size': 'log.append_batch_size',
+    'log_flushes': 'log.flushes',
     'log_rotation_events': 'log.rotation_events',
     'log_rotation_total_time': 'log.rotation_total_time',
     'log_rotation_duration': 'log.rotation_duration',
-    # page cache metrics
-    'page_cache_eviction_exceptions': 'page_cache.eviction_exceptions',
-    'page_cache_evictions': 'page_cache.evictions',
-    'page_cache_flushes': 'page_cache.flushes',
-    'page_cache_hits': 'page_cache.hits',
-    'page_cache_page_faults': 'page_cache.page_faults',
-    'page_cache_pins': 'page_cache.pins',
-    'page_cache_unpins': 'page_cache.unpins',
-    # server metrics
-    'server_threads_jetty_all': 'server.threads.jetty.all',
-    'server_threads_jetty_idle': 'server.threads.jetty.idle',
-    # transaction metrics
+    # database transaction metrics
     'transaction_active': 'transaction.active',
     'transaction_active_read': 'transaction.active_read',
     'transaction_active_write': 'transaction.active_write',
-    'transaction_committed_read': 'transaction.committed_read',
     'transaction_committed': 'transaction.committed',
+    'transaction_committed_read': 'transaction.committed_read',
     'transaction_committed_write': 'transaction.committed_write',
     'transaction_last_closed_tx_id': 'transaction.last_closed_tx_id',
     'transaction_last_committed_tx_id': 'transaction.last_committed_tx_id',
     'transaction_peak_concurrent': 'transaction.peak_concurrent',
-    'transaction_rollbacks_read': 'transaction.rollbacks_read',
     'transaction_rollbacks': 'transaction.rollbacks',
+    'transaction_rollbacks_read': 'transaction.rollbacks_read',
     'transaction_rollbacks_write': 'transaction.rollbacks_write',
     'transaction_started': 'transaction.started',
-    'transaction_terminated_read': 'transaction.terminated_read',
     'transaction_terminated': 'transaction.terminated',
+    'transaction_terminated_read': 'transaction.terminated_read',
     'transaction_terminated_write': 'transaction.terminated_write',
     'transaction_tx_size_heap': 'transaction.tx_size_heap',
     'transaction_tx_size_native': 'transaction.tx_size_native',
+    # page cache metrics
+    'page_cache_bytes_written': 'page_cache.bytes_written',
+    'page_cache_bytes_read': 'page_cache.bytes_read',
+    'page_cache_cancelled_faults': 'page_cache.cancelled_faults',
+    'page_cache_eviction_exceptions': 'page_cache.eviction_exceptions',
+    'page_cache_evictions': 'page_cache.evictions',
+    'page_cache_evictions_cooperative': 'page_cache.evictions_cooperative',
+    'page_cache_flushes': 'page_cache.flushes',
+    'page_cache_hits': 'page_cache.hits',
+    'page_cache_hit_ratio': 'page_cache.hit_ratios',
+    'page_cache_merges': 'page_cache.merges',
+    'page_cache_page_faults': 'page_cache.page_faults',
+    'page_cache_page_fault_failures': 'page_cache.page_fault_failures',
+    'page_cache_pages_copied': 'page_cache.pages_copied',
+    'page_cache_pins': 'page_cache.pins',
+    'page_cache_throttled_times': 'page_cache.throttled_times',
+    'page_cache_throttled_millis': 'page_cache.throttled_millis',
+    'page_cache_unpins': 'page_cache.unpins',
+    'page_cache_usage_ratio': 'page_cache.usage_ratio',
+    # query metrics
+    'db_query_execution_success': 'db_query.execution_success',
+    'db_query_execution_failure': 'db_query.execution_failure',
+    'db_query_execution_latency_millis': 'db_query.execution_latency_millis',
+    # server metrics
+    'server_threads_jetty_all': 'server.threads.jetty.all',
+    'server_threads_jetty_idle': 'server.threads.jetty.idle',
     # JVM GC metrics
     'vm_gc_count_g1_old_generation': 'vm.gc.count.g1_old_generation',
     'vm_gc_count_g1_young_generation': 'vm.gc.count.g1_young_generation',
@@ -222,4 +248,5 @@ METRIC_MAP = {
     # JVM threads metrics
     'vm_thread_count': 'vm.thread.count',
     'vm_thread_total': 'vm.thread.total',
+    'vm_threads': 'vm.threads',
 }